test: modularize run-query so we can test it
