+++ src/progly/run-query.cc
@@ -1,0 +1,351 @@
+/*
+* Copyright (C) 2018 The Regents of the University of California
+* All Rights Reserved
+*
+* This library can redistribute it and/or modify under the terms
+* of the GNU Lesser General Public License Version 2.1 as published
+* by the Free Software Foundation.
+*
+*/
+
+#include <fstream>
+#include <boost/program_options.hpp>
+#include "query.h"
+
+namespace po = boost::program_options;
+
+int main(int argc, char **argv)
+{
+  std::string pool;
+  unsigned num_objs;
+  int wthreads;
+  bool build_index;
+  std::string logfile;
+  int qdepth;
+  std::string dir;
+  std::string projected_col_names_default = "*";
+  std::string projected_col_names;  // provided by the client/user or default
+  Tables::schema_vec current_schema;  // current schema of the table
+  bool apply_predicates = false;  // TODO
+
+  // TODO: get actual table name and schema from the db client
+  int ret = getSchemaFromSchemaString(current_schema, Tables::lineitem_test_schema_string);
+
+  po::options_description gen_opts("General options");
+  gen_opts.add_options()
+    ("help,h", "show help message")
+    ("pool", po::value<std::string>(&pool)->required(), "pool")
+    ("num-objs", po::value<unsigned>(&num_objs)->required(), "num objects")
+    ("use-cls", po::bool_switch(&use_cls)->default_value(false), "use cls")
+    ("quiet,q", po::bool_switch(&quiet)->default_value(false), "quiet")
+    ("query", po::value<std::string>(&query)->required(), "query name")
+    ("wthreads", po::value<int>(&wthreads)->default_value(1), "num threads")
+    ("qdepth", po::value<int>(&qdepth)->default_value(1), "queue depth")
+    ("build-index", po::bool_switch(&build_index)->default_value(false), "build index")
+    ("use-index", po::bool_switch(&use_index)->default_value(false), "use index")
+    ("projection", po::bool_switch(&projection)->default_value(false), "projection")
+    ("build-index-batch-size", po::value<uint32_t>(&build_index_batch_size)->default_value(1000), "build index batch size")
+    ("extra-row-cost", po::value<uint64_t>(&extra_row_cost)->default_value(0), "extra row cost")
+    ("log-file", po::value<std::string>(&logfile)->default_value(""), "log file")
+    ("dir", po::value<std::string>(&dir)->default_value("fwd"), "direction")
+    // query parameters
+    ("extended-price", po::value<double>(&extended_price)->default_value(0.0), "extended price")
+    ("order-key", po::value<int>(&order_key)->default_value(0.0), "order key")
+    ("line-number", po::value<int>(&line_number)->default_value(0.0), "line number")
+    ("ship-date-low", po::value<int>(&ship_date_low)->default_value(-9999), "ship date low")
+    ("ship-date-high", po::value<int>(&ship_date_high)->default_value(-9999), "ship date high")
+    ("discount-low", po::value<double>(&discount_low)->default_value(-9999.0), "discount low")
+    ("discount-high", po::value<double>(&discount_high)->default_value(-9999.0), "discount high")
+    ("quantity", po::value<double>(&quantity)->default_value(0.0), "quantity")
+    ("comment_regex", po::value<std::string>(&comment_regex)->default_value(""), "comment_regex")
+    ("project-col-names", po::value<std::string>(&projected_col_names)->default_value(projected_col_names_default), "projected col names, as csv list")
+  ;
+
+  po::options_description all_opts("Allowed options");
+  all_opts.add(gen_opts);
+
+  po::variables_map vm;
+  po::store(po::parse_command_line(argc, argv, all_opts), vm);
+
+  if (vm.count("help")) {
+    std::cout << all_opts << std::endl;
+    return 1;
+  }
+
+  po::notify(vm);
+
+  assert(num_objs > 0);
+  assert(wthreads > 0);
+  assert(qdepth > 0);
+
+  // connect to rados
+  librados::Rados cluster;
+  cluster.init(NULL);
+  cluster.conf_read_file(NULL);
+  ret = cluster.connect();
+  checkret(ret, 0);
+
+  // open pool
+  librados::IoCtx ioctx;
+  ret = cluster.ioctx_create(pool.c_str(), ioctx);
+  checkret(ret, 0);
+
+  timings.reserve(num_objs);
+
+  // generate the names of the objects to process
+  for (unsigned oidx = 0; oidx < num_objs; oidx++) {
+    std::stringstream oid_ss;
+    oid_ss << "obj." << oidx;
+    const std::string oid = oid_ss.str();
+    target_objects.push_back(oid);
+  }
+
+  if (dir == "fwd") {
+    std::reverse(std::begin(target_objects),
+        std::end(target_objects));
+  } else if (dir == "bwd") {
+    // initial order
+  } else if (dir == "rnd") {
+    std::random_shuffle(std::begin(target_objects),
+        std::end(target_objects));
+  } else {
+    assert(0);
+  }
+
+  // build index for query "d"
+  if (build_index) {
+    std::vector<std::thread> threads;
+    for (int i = 0; i < wthreads; i++) {
+      auto ioctx = new librados::IoCtx;
+      int ret = cluster.ioctx_create(pool.c_str(), *ioctx);
+      checkret(ret, 0);
+      threads.push_back(std::thread(worker_build_index, ioctx));
+    }
+
+    for (auto& thread : threads) {
+      thread.join();
+    }
+
+    return 0;
+  }
+
+  /*
+   * sanity check queries against provided parameters
+   */
+  if (query == "a") {
+
+    assert(!use_index); // not supported
+    assert(extended_price != 0.0);
+    std::cout << "select count(*) from lineitem where l_extendedprice > "
+      << extended_price << std::endl;
+
+  } else if (query == "b") {
+
+    assert(!use_index); // not supported
+    assert(extended_price != 0.0);
+    std::cout << "select * from lineitem where l_extendedprice > "
+      << extended_price << std::endl;
+
+  } else if (query == "c") {
+
+    assert(!use_index); // not supported
+    assert(extended_price != 0.0);
+    std::cout << "select * from lineitem where l_extendedprice = "
+      << extended_price << std::endl;
+
+  } else if (query == "d") {
+
+    if (use_index)
+      assert(use_cls);
+
+    assert(order_key != 0);
+    assert(line_number != 0);
+    std::cout << "select * from from lineitem where l_orderkey = "
+      << order_key << " and l_linenumber = " << line_number << std::endl;
+
+  } else if (query == "e") {
+
+    assert(!use_index); // not supported
+    assert(ship_date_low != -9999);
+    assert(ship_date_high != -9999);
+    assert(discount_low != -9999.0);
+    assert(discount_high != -9999.0);
+    assert(quantity != 0.0);
+    std::cout << "select * from lineitem where l_shipdate >= "
+      << ship_date_low << " and l_shipdate < " << ship_date_high
+      << " and l_discount > " << discount_low << " and l_discount < "
+      << discount_high << " and l_quantity < " << quantity << std::endl;
+
+  } else if (query == "f") {
+
+    assert(!use_index); // not supported
+    assert(comment_regex != "");
+    std::cout << "select * from lineitem where l_comment ilike '%"
+      << comment_regex << "%'" << std::endl;
+
+  } else if (query == "fastpath") {   // no processing required
+
+    assert(!use_index); // not supported
+    assert(!projection); // not supported
+    std::cout << "select * from lineitem" << std::endl;
+
+  } else if (query == "flatbuf") {   // no processing required
+
+    // the queryop schema string will be either the full current schema or projected schema.
+    // set the query schema and check if proj&select
+
+    Tables::schema_vec query_schema;
+    boost::trim(projected_col_names);
+
+    if (projected_col_names == projected_col_names_default) {
+
+        // the query schema is identical to the current schema
+        for (auto it=current_schema.begin(); it!=current_schema.end(); ++it)
+            query_schema.push_back(*it);
+
+        // treat as fastpath query, only if no project and no select
+        if (!apply_predicates)
+            fastpath = true;
+
+    } else {
+        projection = true;
+        Tables::getSchemaFromProjectCols(query_schema, current_schema, projected_col_names);
+        assert(query_schema.size() != 0);
+    }
+
+    table_schema_str = getSchemaStrFromSchema(current_schema);
+    query_schema_str = getSchemaStrFromSchema(query_schema);
+
+    std::cout << "select " << projected_col_names << " from lineitem" << std::endl;
+    cout << "table_schema_str=\n" << table_schema_str << endl;
+    cout << "query_schema_str=\n" << query_schema_str << endl;
+
+  } else {
+    std::cerr << "invalid query: " << query << std::endl;
+    exit(1);
+  }
+
+  result_count = 0;
+  rows_returned = 0;
+  nrows_processed = 0;
+  fastpath |= false;
+
+  outstanding_ios = 0;
+  stop = false;
+
+  // start worker threads
+  std::vector<std::thread> threads;
+  for (int i = 0; i < wthreads; i++) {
+    threads.push_back(std::thread(worker));
+  }
+
+  std::unique_lock<std::mutex> lock(dispatch_lock);
+  while (true) {
+    while (outstanding_ios < qdepth) {
+      // get an object to process
+      if (target_objects.empty())
+        break;
+      std::string oid = target_objects.back();
+      target_objects.pop_back();
+      lock.unlock();
+
+      // dispatch an io request
+      AioState *s = new AioState;
+      s->c = librados::Rados::aio_create_completion(
+          s, NULL, handle_cb);
+
+      memset(&s->times, 0, sizeof(s->times));
+      s->times.dispatch = getns();
+
+      if (use_cls) {
+        query_op op;
+        op.query = query;
+        op.extended_price = extended_price;
+        op.order_key = order_key;
+        op.line_number = line_number;
+        op.ship_date_low = ship_date_low;
+        op.ship_date_high = ship_date_high;
+        op.discount_low = discount_low;
+        op.discount_high = discount_high;
+        op.quantity = quantity;
+        op.comment_regex = comment_regex;
+        op.use_index = use_index;
+        op.projection = projection;
+        op.fastpath = fastpath;
+
+        // this is set above during user input err checking
+        op.table_schema_str = table_schema_str;
+        op.query_schema_str = query_schema_str;
+
+        op.extra_row_cost = extra_row_cost;
+        ceph::bufferlist inbl;
+        ::encode(op, inbl);
+        int ret = ioctx.aio_exec(oid, s->c,
+            "tabular", "query_op", inbl, &s->bl);
+        checkret(ret, 0);
+      } else {
+        int ret = ioctx.aio_read(oid, s->c, &s->bl, 0, 0);
+        checkret(ret, 0);
+      }
+
+      lock.lock();
+      outstanding_ios++;
+    }
+    if (target_objects.empty())
+      break;
+    dispatch_cond.wait(lock);
+  }
+  lock.unlock();
+
+  // drain any still-in-flight operations
+  while (true) {
+    lock.lock();
+    if (outstanding_ios == 0) {
+      lock.unlock();
+      break;
+    }
+    lock.unlock();
+    std::cout << "draining ios: " << outstanding_ios << " remaining" << std::endl;
+    sleep(1);
+  }
+
+  // wait for all the workers to stop
+  work_lock.lock();
+  stop = true;
+  work_lock.unlock();
+  work_cond.notify_all();
+
+  // the threads will exit when all the objects are processed
+  for (auto& thread : threads) {
+    thread.join();
+  }
+
+  ioctx.close();
+
+  if (query == "a" && use_cls) {
+    std::cout << "total result row count: " << result_count
+      << " / -1" << "; nrows_processed=" << nrows_processed
+      << std::endl;
+  } else {
+    std::cout << "total result row count: " << result_count
+      << " / " << rows_returned  << "; nrows_processed=" << nrows_processed
+      << std::endl;
+  }
+
+  if (logfile.length()) {
+    std::ofstream out;
+    out.open(logfile, std::ios::trunc);
+    out << "dispatch,response,read_ns,eval_ns,eval2_ns" << std::endl;
+    for (const auto& time : timings) {
+      out <<
+        time.dispatch << "," <<
+        time.response << "," <<
+        time.read_ns << "," <<
+        time.eval_ns << "," <<
+        time.eval2_ns << std::endl;
+    }
+    out.close();
+  }
+
+  return 0;
+}
